+++ compiler/tests-compiler/util/util.ml
@@ -266,6 +266,7 @@ let extract_sourcemap file =
 let compile_to_javascript
     ?(flags = [])
     ?(use_js_string = false)
+    ?(lambda_lift_all = false)
     ?(effects = `Disabled)
     ?(werror = true)
     ~pretty
@@ -283,6 +284,9 @@ let compile_to_javascript
       ; (if use_js_string
          then [ "--enable=use-js-string" ]
          else [ "--disable=use-js-string" ])
+      ; (if lambda_lift_all
+         then [ "--enable=lambda-lift-all" ]
+         else [ "--disable=lambda-lift-all" ])
       ; flags
       ; (if werror then [ "--Werror" ] else [])
       ]
@@ -324,17 +328,26 @@ let compile_bc_to_javascript
     ?flags
     ?effects
     ?use_js_string
+    ?lambda_lift_all
     ?(pretty = true)
     ?(sourcemap = true)
     ?werror
     file =
   Filetype.path_of_bc_file file
-  |> compile_to_javascript ?flags ?effects ?use_js_string ?werror ~pretty ~sourcemap
+  |> compile_to_javascript
+       ?flags
+       ?effects
+       ?use_js_string
+       ?lambda_lift_all
+       ?werror
+       ~pretty
+       ~sourcemap
 
 let compile_cmo_to_javascript
     ?(flags = [])
     ?effects
     ?use_js_string
+    ?lambda_lift_all
     ?(pretty = true)
     ?(sourcemap = true)
     ?werror
@@ -343,6 +356,7 @@ let compile_cmo_to_javascript
   |> compile_to_javascript
        ?effects
        ?use_js_string
+       ?lambda_lift_all
        ?werror
        ~flags:([ "--disable"; "header" ] @ flags)
        ~pretty
@@ -578,6 +592,7 @@ let compile_and_parse_whole_program
     ?flags
     ?effects
     ?use_js_string
+    ?lambda_lift_all
     ?unix
     ?werror
     s =
@@ -591,11 +606,20 @@ let compile_and_parse_whole_program
            ?flags
            ?effects
            ?use_js_string
+           ?lambda_lift_all
            ?werror
            ~sourcemap:debug
       |> parse_js)
 
-let compile_and_parse ?(debug = true) ?pretty ?flags ?effects ?use_js_string ?werror s =
+let compile_and_parse
+    ?(debug = true)
+    ?pretty
+    ?flags
+    ?effects
+    ?use_js_string
+    ?lambda_lift_all
+    ?werror
+    s =
   with_temp_dir ~f:(fun () ->
       s
       |> Filetype.ocaml_text_of_string
@@ -606,6 +630,7 @@ let compile_and_parse ?(debug = true) ?pretty ?flags ?effects ?use_js_string ?we
            ?flags
            ?effects
            ?use_js_string
+           ?lambda_lift_all
            ?werror
            ~sourcemap:debug
       |> parse_js)